feat: allow passing a device-id as a param on Windows
PS1Transfer currently just defaults to device-id zero. This means that two PS1USB devices cannot be used simultaneously. I have added the option to pass a new param to the application using "-z". The new param is just the index that Windows has assigned to the device. I have not made any changes for Linux, so this param is only currently available when running on Windows.